% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_one_table.R
\name{wilcox_one_table}
\alias{wilcox_one_table}
\title{Function to generate a table to Conduct Wilcoxon Test on}
\usage{
wilcox_one_table(table_length = 1000, sample_from = rpois, ...,
  alpha = 0.05)
}
\arguments{
\item{table_length}{}

\item{sample_from}{}

\item{...}{}

\item{alpha}{}
}
\description{
Function to generate a table to Conduct Wilcoxon Test on
}
